[#89] Split the blacklist method
